requirements.txt
Django==5.0.3
djangorestframework==3.15.1
django-cors-headers==4.3.0
mysqlclient==2.2.4
python-dotenv==1.0.1
